fix(frenchies-spot-api): change type lat lng to FLOAT

+++ prisma/migrations/20230127174909_init/migration.sql
@@ -64,8 +64,8 @@ CREATE TABLE "Spot" (
     "isCanVisit" BOOLEAN NOT NULL DEFAULT false,
     "isTouristic" BOOLEAN NOT NULL DEFAULT false,
     "profileId" TEXT NOT NULL,
-    "lat" INTEGER NOT NULL,
-    "lng" INTEGER NOT NULL,
+    "lat" DOUBLE PRECISION NOT NULL,
+    "lng" DOUBLE PRECISION NOT NULL,
     "region" TEXT NOT NULL,
 
     CONSTRAINT "Spot_pkey" PRIMARY KEY ("id")
